lm35 ok

+++ components/queue/queue.c
@@ -16,7 +16,7 @@ void queue_delete(queue_double_t *queue){
 }
 
 void queue_add(queue_double_t *queue, double value){
-    for(size_t i = queue->size - 1; i >= 0; i--){
+    for(int32_t i = queue->size - 1; i >= 0; i--){
         if(i != 0)
             queue->array[i] = queue->array[i - 1];
     }
@@ -29,7 +29,7 @@ void queue_add(queue_double_t *queue, double value){
 
 double queue_average(queue_double_t *queue){
     double avr = 0;
-    for(size_t i = queue->elements - 1; i >= 0; i--){
+    for(int32_t i = queue->elements - 1; i >= 0; i--){
         avr += queue->array[i];
     }
